requirements.txt
streamlit
pandas
groq
semantic-router
chromadb
python-dotenv
sentence-transformers
